requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
beautifulsoup4==4.12.3
bcrypt~=4.1.2
blinker==1.7.0
cachelib==0.12.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Session==0.8.0
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.13
langchain-openai==0.1.15
langsmith==0.1.85
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msgspec==0.18.6
openai==1.35.13
orjson==3.10.6
packaging==24.1
parso==0.8.4
pdf2image==1.17.0
pdfrw==0.4
pillow==10.4.0
prompt_toolkit==3.0.47
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
traitlets==5.14.3
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.2
zipp==3.18.1
